Move retry/catch to a mixin
