+++ lib/floe.rb
@@ -34,6 +34,7 @@ require_relative "floe/workflow/states/map"
 require_relative "floe/workflow/states/non_terminal_mixin"
 require_relative "floe/workflow/states/parallel"
 require_relative "floe/workflow/states/pass"
+require_relative "floe/workflow/states/retry_catch_mixin"
 require_relative "floe/workflow/states/succeed"
 require_relative "floe/workflow/states/task"
 require_relative "floe/workflow/states/wait"
